Added support for re-spawn locations in CreatureManager. Needs to be instantiated in BatsEverywhere. Currently using a queue data structure, so if a creature dies it should push its original spawn location. Creature Manager will then pop of its top location and use that to create a new creature at that spawn location. Currently only supports Robot but also needs to support mummy and pacman ghost.

+++ src/creatures/Robot.java
@@ -36,7 +36,7 @@ public class Robot implements Creature, PlayerMotionWatcher, ProjectileWatcher{
 	private Location spawnLocation;
 	
 	public Robot(double startX,double startZ,GL2 gl, GLU glu){
-		this.spawnLocation = new Location(startX,0,startZ);
+		this.spawnLocation = new Location((float) startX,(float) 0,(float) startZ);
 		this.robotLocationX = startX;
 		this.robotLocationZ = startZ;
 		eyeVectorX = 0.2;
